moved utils up one directory and adjusted code and action yaml accordingly
