+++ cerulean_cloud/cloud_run_orchestrator/handler.py
@@ -35,12 +35,12 @@ from cerulean_cloud.cloud_run_orchestrator.schema import (
     OrchestratorInput,
     OrchestratorResult,
 )
-from cerulean_cloud.cloud_run_orchestrator.utils import structured_log
 from cerulean_cloud.database_client import DatabaseClient, get_engine
 from cerulean_cloud.models import get_model
 from cerulean_cloud.roda_sentinelhub_client import RodaSentinelHubClient
 from cerulean_cloud.tiling import TMS, offset_bounds_from_base_tiles
 from cerulean_cloud.titiler_client import TitilerClient
+from cerulean_cloud.utils import structured_log
 
 # Configure logger
 logger = logging.getLogger("orchestrate")
