+++ cerulean_cloud/utils.py
